33 improve climate variables interface

+++ src/repicea/simulation/covariateproviders/plotlevel/GrowthStepLengthYrProvider.java
@@ -27,8 +27,8 @@ public interface GrowthStepLengthYrProvider {
 	
 	/**
 	 * This method returns the growth step length in years.
-	 * @return a double
+	 * @return the step length (yrs)
 	 */
-	public double getGrowthStepLengthYr();
+	public int getGrowthStepLengthYr();
 
 }
